full4/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
onnx==1.16.1
onnxruntime==1.19.0

